handler courses partial ready

+++ profileService/internal/domain/staff/service.go
@@ -3,6 +3,6 @@ package staff
 import "context"
 
 type Service interface {
-	GetAllStaffByInstanceID(ctx context.Context, instanceID int) ([]*Staff, error)
+	GetAllStaffByInstanceID(ctx context.Context, instanceID int64) ([]*Staff, error)
 	AddStaff(ctx context.Context, staff *Staff) error
 }
